+++ src/main/java/com/cod/AniBirth/member/controller/MemberController.java
@@ -1,22 +1,20 @@
 package com.cod.AniBirth.member.controller;
 
-import com.cod.AniBirth.account.entity.Account;
-import com.cod.AniBirth.account.service.AccountService;
 import com.cod.AniBirth.email.service.EmailService;
-import com.cod.AniBirth.global.security.DataNotFoundException;
 import com.cod.AniBirth.member.entity.Member;
 import com.cod.AniBirth.member.form.MemberForm;
 import com.cod.AniBirth.member.repository.MemberRepository;
 import com.cod.AniBirth.member.service.MemberService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
@@ -24,9 +22,6 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
-import java.security.Principal;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.UUID;
 
 @Controller
@@ -37,7 +32,6 @@ public class MemberController {
     private final EmailService emailService;
     private final MemberRepository memberRepository;
     private final PasswordEncoder passwordEncoder;
-    private final AccountService accountService;
 
     @PreAuthorize("isAnonymous()")
     @GetMapping("/login")
@@ -180,17 +174,5 @@ public class MemberController {
 
         // 이메일 주소가 일치하지 않거나 회원이 존재하지 않는 경우
         return "redirect:/member/login?incorrect=true";
-    }
-
-    @PreAuthorize("isAuthenticated()")
-    @GetMapping("/myPage")
-    public String myPage(Model model, Principal principal) {
-        Member member = memberService.findByUsername(principal.getName());
-        Account account = accountService.findByMember(member);
-
-        model.addAttribute("member", member);
-        model.addAttribute("account", account);
-
-        return "member/myPage";
     }
 }
