+++ src/main/resources/templates/layout/layout.html
@@ -34,12 +34,15 @@
     <link href="/resource/common/common.css" rel="stylesheet" type="text/css"/>
     <link href="/resource/home/main.css" rel="stylesheet" type="text/css"/>
     <link href="/resource/member/signup.css" rel="stylesheet" type="text/css"/>
-    <link href="/resource/member/myPage.css" rel="stylesheet" type="text/css"/>
+    <link href="/resource/member/myPage/myProfile.css" rel="stylesheet" type="text/css"/>
+    <link href="/resource/member/myPage/volunteer.css" rel="stylesheet" type="text/css"/>
+    <link href="/resource/member/myPage/adopt.css" rel="stylesheet" type="text/css"/>
+    <link href="/resource/member/myPage/donation.css" rel="stylesheet" type="text/css"/>
 
     <!-- 커스텀 js모음 -->
     <script defer src="/javascript/home/main.js"></script>
     <script defer src="/javascript/member/signup.js"></script>
-    <script defer src="/javascript/member/myPage.js"></script>
+    <script defer src="/javascript/member/myPage/myProfile.js"></script>
 </head>
 
 <body class="flex flex-col min-h-screen pt-[56px]">
@@ -60,7 +63,7 @@
                         <span sec:authorize="isAuthenticated()" th:text="${#strings.startsWith(#authentication.name, 'NAVER') ? #authentication.name.substring(0, 20) : #authentication.name} + '님'"></span>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" sec:authorize="isAuthenticated()" th:href="@{/member/myPage}">마이페이지</a>
+                        <a class="nav-link" sec:authorize="isAuthenticated()" th:href="@{/member/myPage/myProfile}">마이페이지</a>
                     </li>
                     <li class="nav-item">
                         <a class="nav-link" sec:authorize="isAuthenticated()" href="#">장바구니</a>
